requirements.txt
certifi==2019.11.28
paho-mqtt==1.5.0
dash==2.0.0
plotly==5.3.1
pandas==1.3.3